Remove excess experiments
