--- time_representations/hist_data_analysis/interpolation/mean_stds.json
@@ -1,1 +1,0 @@
-{"humidity": [61.251041608910924, 15.18131083057179], "pressure": [1012.9782961044582, 5.365216536322366], "feels_like": [17.97240415733361, 9.181487919382173], "temp": [18.63217415848606, 8.268436920458482], "wind_speed": [4.561839727750256, 2.552532578558552], "rain_1h": [0.1445253348864298, 1.5549978710191492]}
