+++ time_representations_hourly/config.yaml
@@ -1,4 +1,4 @@
-seeds: [72, 834, 1213]
+seeds: [8, 72, 265, 834, 1013, 2910]
 bins: [
         ["ValueRangeBin",  [0.7049093788491994, 0.11666373394333979,  0.048389934893542146, 0.032729192327995774,  0.09730775998592293]]
         ]
